Split router.rs test module out to stay under the file-size cap
The TAN-544/TAN-548 unit tests pushed incident_monitor/router.rs to 2081
lines, over the Lint Backend 2000-line hard cap. Move the #[cfg(test)] mod
tests body into router_tests.rs referenced via #[path]; router.rs drops to
1491 lines. No test changes.
